added the entry code options in the example overlay and moved the example verification to the example overlay page
